Fixing a 32bit overflow issue and an hlsl compiler warning (#385)
Fixing a 32bit issue and making a unittest that verifies the fix. Also fixing an implicit truncation warning that was being spewed from the HLSL compiler

+++ Libraries/D3D12RaytracingFallback/src/GpuBvh2Builder.h
@@ -51,17 +51,17 @@ namespace FallbackLayer
 
         struct ScratchMemoryPartitions
         {
-            UINT OffsetToSceneAABB;
-            UINT OffsetToElements;
-            UINT OffsetToMortonCodes;
-            UINT OffsetToIndexBuffer;
-            UINT OffsetToHierarchy;
+            UINT64 OffsetToSceneAABB;
+            UINT64 OffsetToElements;
+            UINT64 OffsetToMortonCodes;
+            UINT64 OffsetToIndexBuffer;
+            UINT64 OffsetToHierarchy;
 
-            UINT OffsetToSceneAABBScratchMemory;
+            UINT64 OffsetToSceneAABBScratchMemory;
 
-            UINT OffsetToCalculateAABBDispatchArgs;
-            UINT OffsetToPerNodeCounter;
-            UINT TotalSize;
+            UINT64 OffsetToCalculateAABBDispatchArgs;
+            UINT64 OffsetToPerNodeCounter;
+            UINT64 TotalSize;
         };
 
         ScratchMemoryPartitions CalculateScratchMemoryUsage(Level level, UINT numTriangles);
